requirements.txt
altair==4.2.2
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
branca==0.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
deprecation==2.1.0
entrypoints==0.4
exceptiongroup==1.1.1
executing==1.2.0
Flask==2.3.2
folium==0.14.0
fonttools==4.39.3
functions-framework==3.3.0
gcloud==0.17.0
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-auth==2.18.0
google-cloud-core==2.3.2
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
gunicorn==20.1.0
h11==0.14.0
html5lib==1.1
httplib2==0.22.0
idna==3.4
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.13.2
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonschema==4.17.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-tabnine==0.0.24
jws==0.1.3
kiwisolver==1.4.4
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
nest-asyncio==1.5.6
numpy==1.24.3
oauth2client==3.0.0
outcome==1.2.0
packaging==23.1
pandas==2.0.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycryptodome==3.4.3
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-jwt==2.0.1
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.0.2
requests==2.30.0
requests-toolbelt==0.7.0
rich==13.3.5
rsa==4.9
seaborn==0.12.2
selenium==4.9.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
stack-data==0.6.2
streamlit==1.22.0
streamlit-folium==0.11.1
streamlit-lottie==0.0.3
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.1
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
urllib3==1.26.15
validators==0.20.0
watchdog==3.0.0
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.4
wsproto==1.2.0
zipp==3.15.0
